third_party/blink/web_tests/fast/lists/li-values-expected.txt

1 The list marker should show 1 because this item had a value of 4 but it was removed.
-1 The list marker should show -1 (negative values are allowed).
0 The list marker should show 0.
1 The list marker should show 1 (invalid values are ignored).
-5 The list marker should show -5 (updated programmatically).
마이너스 일 The list marker should show -1 (negative values are allowed).
-1 The list marker should fall back to decimal -1 (hebrew: 1-infinite).
0 The list marker should fall back to decimal 0 (simple-upper-roman: 1-4999).
0 The list marker should fall back to decimal 0 (simple-lower-roman: 1-4999).
0 The list marker should fall back to decimal 0 (upper-roman: 1-4999).
0 The list marker should fall back to decimal 0 (lower-roman: 1-4999).
0 The list marker should fall back to decimal 0 (lower-armenian: 1-9999).
0 The list marker should fall back to decimal 0 (upper-armenian: 1-9999).
0 The list marker should fall back to decimal 0 (armenian: 1-9999).
0 The list marker should fall back to decimal 0 (georgian: 1-19999).
0 The list marker should fall back to decimal 0 (ancient-tamil: 1-9999).
マイナス一 The list marker should fall back to cjk-decimal -1 (japanese-informal: 0-9999).
マイナス壱 The list marker should fall back to cjk-decimal -1 (japanese-formal: 0-9999).
마이너스 一 The list marker should fall back to decimal -1 (korean-hanja-informal: 0-9999).
마이너스 壹 The list marker should fall back to decimal -1 (korean-hanja-formal: 0-9999).
0 The list marker should fall back to decimal 0 (greek: 1-999).
